local_requirements.txt
streamlit>=1.28.0
crewai>=0.28.8
ollama>=0.1.7
requests>=2.31.0
python-dotenv>=1.0.0
pathlib
ddgs>=9.0.0
reportlab>=4.0.0
xhtml2pdf>=0.2.17
